# auth/src/swagger/swagger.yml
openapi: 3.0.0
info:
  title: Software2 API Documentation
  description: |
    This is the API documentation for the Software2 project. It provides details about the available authentication, permission, and role endpoints, their usage, and expected inputs/outputs.
  termsOfService: https://example.com/terms/
  contact:
    email: support@software2.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.software2.com
    description: Production server
...
paths:
...
  /permissions:
    post:
      tags:
        - permission
      summary: Create a new permission
      description: Add a new permission to the system.
      operationId: createPermission
      requestBody:
        description: Permission object to be created
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "View Users"
                url:
                  type: string
                  example: "/api/users"
                method:
                  type: string
                  example: "GET"
                description:
                  type: string
                  example: "Permission to view user details"
      responses:
        '201':
          description: Permission created successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
    get:
      tags:
        - permission
      summary: Get all permissions
      description: Retrieve a list of all permissions in the system.
      operationId: getPermissions
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '500':
          description: Server error
  /permissions/{id}:
    get:
      tags:
        - permission
      summary: Get permission by ID
      description: Retrieve a specific permission by its ID.
      operationId: getPermissionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the permission
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '404':
          description: Permission not found
        '500':
          description: Server error
    put:
      tags:
        - permission
      summary: Update a permission
      description: Update the details of an existing permission.
      operationId: updatePermission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the permission
      requestBody:
        description: Permission object to be updated
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Edit Users"
                url:
                  type: string
                  example: "/api/users/edit"
                method:
                  type: string
                  example: "PUT"
                description:
                  type: string
                  example: "Permission to edit user details"
      responses:
        '200':
          description: Permission updated successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
    delete:
      tags:
        - permission
      summary: Delete a permission
      description: Remove a permission from the system.
      operationId: deletePermission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the permission
      responses:
        '200':
          description: Permission deleted successfully
        '404':
          description: Permission not found
        '500':
          description: Server error
  /roles:
    post:
      tags:
        - role
      summary: Create a new role
      description: Add a new role to the system.
      operationId: createRole
      requestBody:
        description: Role object to be created
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Admin"
                description:
                  type: string
                  example: "Administrator role with full access"
      responses:
        '201':
          description: Role created successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
    get:
      tags:
        - role
      summary: Get all roles
      description: Retrieve a list of all roles in the system.
      operationId: getRoles
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '500':
          description: Server error
  /roles/{id}:
    get:
      tags:
        - role
      summary: Get role by ID
      description: Retrieve a specific role by its ID.
      operationId: getRoleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the role
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: Role not found
        '500':
          description: Server error
    put:
      tags:
        - role
      summary: Update a role
      description: Update the details of an existing role.
      operationId: updateRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the role
      requestBody:
        description: Role object to be updated
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Editor"
                description:
                  type: string
                  example: "Editor role with limited access"
      responses:
        '200':
          description: Role updated successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
    delete:
      tags:
        - role
      summary: Delete a role
      description: Remove a role from the system.
      operationId: deleteRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the role
      responses:
        '200':
          description: Role deleted successfully
        '404':
          description: Role not found
        '500':
          description: Server error
  /roles/{id}/permissions:
    post:
      tags:
        - role
      summary: Add permissions to a role
      description: Add permissions to a specific role.
      operationId: addPermissionsToRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the role
      requestBody:
        description: Permissions to be added
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: array
                  items:
                    type: string
                  example: ["permissionId1", "permissionId2"]
      responses:
        '200':
          description: Permissions added successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
    delete:
      tags:
        - role
      summary: Remove permissions from a role
      description: Remove permissions from a specific role.
      operationId: removePermissionsFromRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the role
      requestBody:
        description: Permissions to be removed
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: array
                  items:
                    type: string
                  example: ["permissionId1", "permissionId2"]
      responses:
        '200':
          description: Permissions removed successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
...
components:
  schemas:
    Permission:
      type: object
      properties:
        id:
          type: string
          example: "abcde"
        name:
          type: string
          example: "View Users"
        url:
          type: string
          example: "/api/users"
        method:
          type: string
          example: "GET"
        description:
          type: string
          example: "Permission to view user details"
    Role:
      type: object
      properties:
        id:
          type: string
          example: "67890"
        name:
          type: string
          example: "Admin"
        description:
          type: string
          example: "Administrator role with full access"
...